[level 0] Title: 배열의 평균값, Time: 0.03 ms, Memory: 74.6 MB -BaekjoonHub

+++ 프로그래머스/lv0/120817. 배열의 평균값/배열의 평균값.java
@@ -5,9 +5,8 @@ class Solution {
 
         for(int i=0;i<numbers.length;i++){
             sum += numbers[i];
-            answer = (double) sum/numbers.length;
+            answer = sum/numbers.length;
         }
-        
         return answer;
     }
 }
